pyproject.toml
[tool.poetry]
name = "django-turbo-helper"
version = "2.1.0"
description = "Hotwire/Turbo helpers for Django, inspired by Rails."
authors = ["Michael Yin <michaelyin@accordbox.com>"]
license = "MIT"
homepage = "https://github.com/rails-inspire-django/django-turbo-helper"
readme = "README.md"
packages = [{ include = "turbo_helper", from = "src" }]

[tool.poetry.urls]
Changelog = "https://github.com/rails-inspire-django/django-turbo-helper/releases"

[tool.poetry.dependencies]
python = ">=3.8"
django = ">=3.0"
django-actioncable = ">=1.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["setuptools", "poetry_core>=1.0"]
build-backend = "poetry.core.masonry.api"
